refactor: remove  core/services folder and move its content to core
